==> notebooks/03 - Vetorização e Salvamento dos Dados.ipynb <==
import rasterio
import geopandas as gpd
import numpy as np
from rasterio.mask import mask
import pandas as pd
import os
from rasterio.features import shapes
from shapely.geometry import shape

# camada corte

distritos_path = "ShapesProjetoFpolisMapBiomas.gpkg"
layer_distritos = "geometrias_corrigidas"

camada_corte = gpd.read_file(distritos_path, layer=layer_distritos)

# camada raster
# vetoriza

gpkg_saida = "RastersVetorizados_v1.gpkg"

for ano in range(1986,2025):
    raster_path = f"{ano}_31982.tif"
    with rasterio.open(raster_path) as src:
        raster_anual = src.read(1)      # banda 1
        transform = src.transform
        crs = src.crs
        nodata = src.nodata

    geometrias = []
    valores = []
    
    for geom, value in shapes(raster_anual, transform=transform):
        if value is not None:
            geometrias.append(shape(geom))
            valores.append(value)
    
        
    # Converter para GeoDataFrame
    gdf = gpd.GeoDataFrame(
        {"valor": valores},
        geometry=geometrias,
        crs=crs
        )

    # Intersecta
    gdf_intersec = gpd.overlay(
        camada_corte,
        gdf,
        how="intersection"
        )

    # ➕ adiciona o ano
    gdf_intersec["ano"] = ano

    # Nome da camada
    nome_camada = f"{ano}_vetorizado"

    # Exporta (append no mesmo GPKG)
    gdf_intersec.to_file(
        gpkg_saida,
        layer=nome_camada,
        driver="GPKG"
    )

    print(f"Camada {nome_camada} exportada com sucesso.")

gdf_intersec.head()
