==> notebooks/04 - Verificação Area.ipynb <==
import fiona

gpkg_path = "RastersVetorizados_v1.gpkg"

layers = fiona.listlayers(gpkg_path)

print("Camadas no GeoPackage:")
for layer in layers:
    print("-", layer)

import geopandas as gpd

gdf = gpd.read_file(gpkg_path, layer="1985_vetorizado")
print(gdf.head())

gdf["area_feicao_m2"] = gdf.geometry.area

print(gdf.head())

import pandas as pd

df_soma = (
    gdf
    .groupby("nome", as_index=False)
    .agg(
        area_calculada_m2=("area_feicao_m2", "sum"),
        area_referencia_m2=("area_m2", "first")
    )
)

df_soma["dif_m2"] = (
    df_soma["area_calculada_m2"] - df_soma["area_referencia_m2"]
)

df_soma["dif_percentual"] = (
    df_soma["dif_m2"] / df_soma["area_referencia_m2"] * 100
)

display(df_soma.sort_values("dif_percentual", ascending=False))

df_soma.to_csv('Diferenças nas áreas.csv')

import geopandas as gpd

gdf1995 = gpd.read_file(gpkg_path, layer="1995_vetorizado")
gdf1995["area_feicao_m2"] = gdf1995.geometry.area
df_1995 = (
    gdf1995
    .groupby("nome", as_index=False)
    .agg(
        area_calculada_m2=("area_feicao_m2", "sum"),
        area_referencia_m2=("area_m2", "first")
    )
)

df_1995["dif_m2"] = (
    df_1995["area_calculada_m2"] - df_1995["area_referencia_m2"]
)

df_1995["dif_percentual"] = (
    df_1995["dif_m2"] / df_1995["area_referencia_m2"] * 100
)

display(df_1995.sort_values("dif_percentual", ascending=False))
