==> notebooks/05 - Camadas para Tabela.ipynb <==
import geopandas as gpd
import pandas as pd
import fiona

gpkg_path = "RastersVetorizados_v1.gpkg"
layers = fiona.listlayers(gpkg_path)

dfs = []

for layer in layers:
    print(f"Processando camada: {layer}")
    
    gdf = gpd.read_file(gpkg_path, layer=layer)
    
    gdf["area_feicao_m2"] = gdf.geometry.area
    
    dfs.append(gdf)

# Unir tudo em um único DataFrame
df_final = pd.concat(dfs, ignore_index=True)

display(df_final.head())


import pandas as pd

legenda = pd.read_csv("legenda_mapbiomas_colecao10.csv", delimiter=';')

df_final["valor"] = df_final["valor"].astype("Int64")  # aceita NaN
legenda["id"] = legenda["id"].astype("Int64")
df_merged = df_final.merge(
    legenda,
    left_on="valor",
    right_on="id",
    how="left"
)

df_merged = df_merged.drop(
    columns=["codigo_ibg", "id_y", "area_km2"],
    errors="ignore"  # evita erro se alguma não existir
)
display(df_merged)

df_merged.to_csv('area_feicoes_todos_anos.csv', sep = ';')

df_merged.to_excel('area_feicoes_todos_anos.xlsx')
